add tables for parking space matching algorithm
prereserved_list, match a waiting user with a parking space, waiting for user
	to confirm the reservation.

waiting_pool, pool that contains all the users that are waiting for a parking space.

parking_space_pool, pool that contains all the parking spaces that are available.

those tables will be refactored in future and move to redis.